+++ SourceryTests/Parsing/AnnotationsParserSpec.swift
@@ -42,7 +42,16 @@ class AnnotationsParserSpec: QuickSpec {
                 }
 
                 it("extracts inline annotations") {
-                    let result = parse("//sourcery: skipDescription\n/* sourcery: skipEquality */var name: Int { return 2 }")
+                    let result = parse("//sourcery: skipDescription\n/* sourcery: skipEquality */\n/** sourcery: skipCoding */var name: Int { return 2 }")
+                    expect(result).to(equal([
+                        "skipDescription": NSNumber(value: true),
+                        "skipEquality": NSNumber(value: true),
+                        "skipCoding": NSNumber(value: true)
+                        ]))
+                }
+
+                it("extracts inline annotations from multi line comments") {
+                    let result = parse("//**\n*Comment\n*sourcery: skipDescription\n*sourcery: skipEquality\n*/var name: Int { return 2 }")
                     expect(result).to(equal([
                         "skipDescription": NSNumber(value: true),
                         "skipEquality": NSNumber(value: true)